5.22 upd8

+++ src/main/java/com/maomingming/tpcc/txn/StockLevelTxn.java
@@ -8,7 +8,7 @@ public class StockLevelTxn {
     public int w_id;
     public int d_id;
     public int threshold;
-    public int low_stock;
+    public long low_stock;
 
     public StockLevelTxn(int w_id, int t_id) {
         this.w_id = w_id;
